fix: print warnings for unresolvable uris
Unresolveable URIs weren't being properly logged.

This is because `logging.warning()` does not work
properly in multiprocessing.

(there's some ways to hack around it with multiprocessing.get_logger(),
 but even then there's issues with that)

Instead, what we're doing now is raising and catching `Warning`s in the
child process, then returning them to the main process to
logging.warn().